Refactor code, implement input & output routes, and create simple homepage
