+++ meeting-app/package.json
@@ -10,6 +10,7 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/kit": "^1.5.0",
+		"carbon-components-svelte": "^0.73.2",
 		"svelte": "^3.54.0",
 		"vite": "^4.0.0"
 	},
