+++ meeting-app/src/app.html
@@ -3,6 +3,10 @@
 	<head>
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
+		<link
+			rel="stylesheet"
+			href="https://unpkg.com/carbon-components-svelte/css/white.css"
+	  	/>
 		<meta name="viewport" content="width=device-width" />
 		%sveltekit.head%
 	</head>
