+++ meeting-app/src/routes/+layout.svelte
@@ -1,5 +1,11 @@
 <script>
-    import '../global.css'
+  import "../global.css";
 </script>
-  
+
+<!-- 
+<nav>
+  <a href="/">Home</a>
+  <a href="/input/">Input</a>
+</nav> -->
+
 <slot />
