+++ meeting-app/src/routes/form.svelte
@@ -1,0 +1,40 @@
+<script>
+  import { goto } from "$app/navigation";
+
+  import { FluidForm, TextInput, Button } from "carbon-components-svelte";
+
+  export let username = "";
+  export let submitted = false;
+
+  const handleSubmit = () => {
+    console.log(username);
+    submitted = true;
+  };
+</script>
+
+<FluidForm
+  on:submit={(e) => {
+    e.preventDefault();
+    console.log("submit", e);
+    handleSubmit();
+    goto("/input/" + username);
+  }}
+>
+  <TextInput
+    bind:value={username}
+    labelText="User name"
+    placeholder="Enter user name..."
+    required
+  />
+  <br />
+  <div class="buttonBox">
+    <Button type="submit">Enter</Button>
+  </div>
+</FluidForm>
+
+<style>
+  .buttonBox {
+    display: flex;
+    justify-content: center;
+  }
+</style>
