+++ meeting-app/src/routes/input/[name]/+page.js
@@ -1,0 +1,7 @@
+export function load({ params }) {
+	const name = params.name;
+
+	return {
+		name
+	};
+}
