+++ meeting-app/src/routes/output/+page.svelte
@@ -1,0 +1,1 @@
+<h2>this is the output!</h2>
